@@ -3,40 +3,40 @@ package fr.dauphine.ja.lamhandyhajar.morpionsolitaire.core;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public class Pair<T, E> {
-	protected T p1;
-	protected E p2;
+    protected T p1;
+    protected E p2;
 
-	public T getP1() {
-		return p1;
-	}
+    /**
+     * @param p1
+     * @param p2
+     * @throws NullPointerException
+     */
+    public Pair(T p1, E p2) {
+        checkNotNull(p1);
+        checkNotNull(p2);
 
-	public E getP2() {
-		return p2;
-	}
+        this.p1 = p1;
+        this.p2 = p2;
+    }
 
-	/**
-	 * @param p1
-	 * @param p2
-	 * @throws NullPointerException
-	 */
-	public Pair(T p1, E p2) {
-		checkNotNull(p1);
-		checkNotNull(p2);
-		
-		this.p1 = p1;
-		this.p2 = p2;
-	}
+    public T getP1() {
+        return p1;
+    }
 
-	@Override
-	public boolean equals(Object o) {
-		if (o == this)
-			return true;
+    public E getP2() {
+        return p2;
+    }
 
-		if (!(o instanceof Pair))
-			return false;
-		Pair<?, ?> p = (Pair<?, ?>) o;
+    @Override
+    public boolean equals(Object o) {
+        if (o == this)
+            return true;
 
-		return p.p1.equals(p1) && p.p2.equals(p2);
-	}
-	
+        if (!(o instanceof Pair))
+            return false;
+        Pair<?, ?> p = (Pair<?, ?>) o;
+
+        return p.p1.equals(p1) && p.p2.equals(p2);
+    }
+
 }
